fix: use full OIDC server metadata as cache key MONGOSH-1434
We probably want to trigger a full re-authentication when the server
metadata does not match the metadata we had used for a previous
authentication event.

Also, clarify a note about the current state of this plugin with
regards to being able to pass it to multiple MongoClient instances.
(This does not affect DevTools as we currently do not have plans
to re-use OIDC plugin instances across different MongoDB endpoints
at the same time, but may be relevant for other users as well as
our future selves.)