feat: do not enable device-auth flow by default
In WRITING-14037, one of the items for driver OIDC callbacks was:

> Recommend considering whether to allow Device Authorization Grant

In mongosh and Compass, we do not want to enable this by default.
To make it easier for other Node.js driver users to also use the
same defaults as us, let’s flip the default here in the plugin.