chore: do slightly stricter URL validation (#92)
The URL that is passed to the browser in our plugin is generated
by us, so we are in control over the format of it.
This commit adds another check that that format is still intact
right before we open the browser, so that e.g. a future developer
who wants to remove our redirect would not be able to accidentally
introduce a security issue by letting URLs pass through which contain
characters with special semantics in the shell (e.g. `&` in a
query string).