chore: add better releasing docs (#204)
This PR adds a `Contributing.md` that has a releasing section and 
updates the PR description of the prepare-release workflow with 
instructions for how to finish the release.

Additionally, it replaces the token used to creating and merging the 
release PR with the token we have for devtoolsbot. This should result
in normal workflows getting run both on the release PR and the main
branch after merging.