fix: improve node-fetch compatibility in error cases MONGOSH-2443
`node-fetch` currently does not return a body that is a web/standard
`ReadableStream`, which is what openid-client expects. Since it only
makes use of this in error cases, no additional cases are broken,
but the error message is highly unhelpful in these cases.

Unfortunately, it is not trivial to resolve this without fully wrapping
the `node-fetch`, so this provides a limited compatibility layer rather
than a full-featured one.