chore(deps): upgrade MongoDB Driver and node-fetch to latest MONGOSH-2978 (#237)
* chore: Upgrade MongoDB Driver and node-fetch to latest

* chore: Fix integration tests with Okta and update documentation

There are many changes on how Okta works now and our tests were broken
because of them. Essentially, we can split the changes in two
sections:

1. In the Auth window, now the username and password are in two
sections that require clicking on Submit first.

2. The application configuration in Okta changed drastically,
forcing now MFA on any default application. I've updated the
okta-setup.md to disable that behaviour on the custom
application used for testing.

* chore: Clarify that you can use Okta Verify in your laptop

Okta Verify is now mandatory by default, but you don't need
to install it in your phone if you already have it in your
laptop.

* chore: Run on the latest macOS distro